# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive REST API for managing game server infrastructure with panel integration, async job queue management, and real-time sync operations",
        "title": "NodeByte API",
        "termsOfService": "https://nodebyte.co.uk/legal/terms",
        "contact": {
            "name": "Contact Support",
            "url": "https://discord.gg/wN58bTzzpW"
        },
        "license": {
            "name": "AGPL 3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "0.2.0"
    },
    "host": "core.nodebyte.host",
    "basePath": "/",
    "paths": {
...
        "/api/v1/hytale/server-logs": {
            "get": {
                "description": "Retrieves stored console output logs from a Hytale game server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hytale Logs"
                ],
                "summary": "Get Hytale game server logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server UUID",
                        "name": "server_uuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum logs to return (default: 100, max: 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server logs retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.ServerLog"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Server not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Stores console output logs sent from Wings daemon or Panel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hytale Logs"
                ],
                "summary": "Create game server logs",
                "parameters": [
                    {
                        "description": "Logs to store",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateServerLogsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Logs stored successfully",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hytale/server-logs/count": {
            "get": {
                "description": "Returns the total number of stored logs for a specific server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hytale Logs"
                ],
                "summary": "Get Hytale server logs count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server UUID",
                        "name": "server_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log count retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.CreateServerLogsRequest": {
            "type": "object",
            "properties": {
                "account_id": {
                    "description": "Account ID of the server owner",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "logs": {
                    "description": "Array of log lines to store",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.LogEntry"
                    }
                },
                "server_uuid": {
                    "description": "Server UUID (for Hytale servers, this is the profile UUID)",
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
...
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Human-readable error message",
                    "type": "string",
                    "example": "account_id is required"
                },
                "success": {
                    "description": "Whether the request was successful",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "types.LogEntry": {
            "type": "object",
            "properties": {
                "line": {
                    "description": "Log line content",
                    "type": "string",
                    "example": "[10:30:45] Server started successfully"
                },
                "timestamp": {
                    "description": "Optional timestamp (defaults to now)",
                    "type": "string",
                    "example": "2026-01-14T10:30:45Z"
                }
            }
        },
...
        "types.ServerLog": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Log entry ID",
                    "type": "integer",
                    "example": 1
                },
                "line": {
                    "description": "Log line content",
                    "type": "string",
                    "example": "[10:30:45] Server started successfully"
                },
                "timestamp": {
                    "description": "When the log was created",
                    "type": "string",
                    "example": "2026-01-14T10:30:45Z"
                }
            }
        },
        "types.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data (generic for various endpoints)"
                },
                "message": {
                    "description": "Optional message",
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "description": "Whether the request was successful",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Backend-to-backend API authentication",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}